FINALLY IT WORKS AND PASSES ALL TESTS!
The issue was the following: I had labels for the possible scenarios
when comparing values (lt, gt, eq) and which value should be stored in
the stack: true or false. Because every time a comparison operation was
done, the labels to the commands to decide to store a True/False all had
the same names, so the instructions always jumped to the last one
defined, losing track of where the next instruction should run. To solve
this, I generated a unique random ID for each label every time a
comparison operator happened, in that way each one had their own unique
set of labels (with repeated instructions lol).

It's hacky, but it works!